+++ src/vue/modules/loyalty-points/create-invoice-form/index.vue
@@ -16,7 +16,7 @@
               :label="'create-invoice-form.amount-lbl' | globalize({
                 asset: form.asset
               })"
-              :step="MINIMAL_NUMBER_INPUT_STEP"
+              :step="MIN_AMOUNT"
               :error-message="getFieldErrorMessage(
                 'form.amount',
                 {
@@ -25,6 +25,20 @@
                 }
               )"
               :disabled="amount || formMixin.isDisabled"
+            />
+          </div>
+        </div>
+
+        <div class="app__form-row">
+          <div class="app__form-field">
+            <input-field
+              white-autofill
+              v-model="form.subject"
+              @blur="touchField('form.subject')"
+              name="create-invoice-subject"
+              :label="'create-invoice-form.subject-lbl' | globalize"
+              :error-message="getFieldErrorMessage('form.subject')"
+              :disabled="subject || formMixin.isDisabled"
             />
           </div>
         </div>
@@ -177,6 +191,9 @@ const QR_CODE_BASE = 'tokend://loyaltypay?url='
 const INVOICE_URL_BASE = 'https://go.tokend.io/loyaltypay?url='
 const POLL_INTERVAL = 5000
 
+const MIN_AMOUNT = 0.01
+const DECIMAL_POINTS = 2
+
 export default {
   name: 'create-invoice-form',
   components: {
@@ -204,15 +221,20 @@ export default {
       required: false,
       default: '',
     },
+    subject: {
+      type: String,
+      required: false,
+      default: '',
+    },
   },
   data: _ => ({
-    MIN_AMOUNT: config.MIN_AMOUNT,
+    MIN_AMOUNT,
     MAX_AMOUNT: config.MAX_AMOUNT,
-    DECIMAL_POINTS: config.DECIMAL_POINTS,
-    MINIMAL_NUMBER_INPUT_STEP: config.MINIMAL_NUMBER_INPUT_STEP,
+    DECIMAL_POINTS,
     isInitialized: false,
     form: {
       amount: '',
+      subject: '',
       merchant: '',
       asset: '',
       account: '',
@@ -229,8 +251,11 @@ export default {
       form: {
         amount: {
           required,
-          minValue: minValue(this.MIN_AMOUNT),
-          maxDecimalDigitsCount: maxDecimalDigitsCount(config.DECIMAL_POINTS),
+          minValue: minValue(MIN_AMOUNT),
+          maxDecimalDigitsCount: maxDecimalDigitsCount(DECIMAL_POINTS),
+        },
+        subject: {
+          required,
         },
       },
     }
@@ -331,6 +356,7 @@ export default {
       this.form.account = this.accountId
       this.form.system = config.HORIZON_SERVER
       this.form.amount = this.amount
+      this.form.subject = this.subject
     },
     initPolling () {
       this.pollIntervalId = setInterval(this.checkForPayment, POLL_INTERVAL)